Support for plotting x axis selection
Added support for arbitrary x-axis selection in the plot section.
If a time column exists it will plot against that by default; otherwise it assumes the default assumes the device has an attribute sampling, with the sampling rate expressed as "# [S/s]", which is employed to create the x coordinates for the plot.