Implemented a first version of socket server-client control
SocketDeviceClient and SocketDeviceServer wrap a driver class for server side and client side communication between different computers (or on the same computer).
Due to the way the main DAQ program works they can't run in the same main instance. This is because some drivers call class methods in __init__which are wrapped to function with sockets, but initially the drivers are just initialized one by one to check if the connection is valid. Then the client tries to connect to the server class in order to execute those functions, but there is no server listening, resulting in a crashing DAQ program.