Fixed bug in InfluxDB enable
Apparently when clicking enable on the GUI for InfluxDB monitoring it pushes it as an int, where 1 == False and 2 == True for some bizarre reason. The code assumes self.parent.config["influxdb"]["enabled"] is a str.
By putting a try except catch for an AttributeError (can't strip() an int) the bug is resolved.
I noticed when clicking the button it looks different vs setting it to True in the .ini file, not sure what causes that.