Changed monitoring_commands fom list to set
When adevice takes longer to evaluate a monitoring function than  the monitoring thread takes to add an monitoring function to the shared monitoring_commands list, the device is stuck in an infinite loop while evaluating monitoring commands. Changing the monitoring_commands list to a set prevents multiple evaluations of the same command, hence solving the infite loop.