allow dummy variables
indicate a dummy variable with $x, where x is the index of the nested loop you want to grab the variable from. x counts down from the top, e.g. the first loop is $1, the second nested loop is $2. If you want to grab the device name, or function called, use $devx or $fnx